Error 5013 on Helpdesk change password (href embedded in userKey)
Problem is with "href" as part of the encrypted userKey -- The data has
been encrypted and in this case contains the key word "href".  The code
to read the data first passes it through a protection subroutine that
checks for specific key words.

Because this data is actually encrypted, there is no need to protect
against specific key words.

The fix is to use a flag to specify that we don't need to do the protection
subroutine.