Updated to the latest version of httpclient (4.5.1)
To safely rework the code to remove the deprecation warnings, I added unit tests which leverage the WireMock HTTP server mocking framework.  By doing this, I was able to capture the functionality from how the old httpclient APIs were being used, and duplicate that functionality using the 4.5.1 APIs.

By having these unit tests, we can keep up with changes to httpclient in the future, and be confident the existing functionality still works.