Update PwmRequest to work with both Tomcat7 and Jetty
HttpServletRequest getRequestURI() on jetty returns the jsp page which the user is forwarded to.
HttpServletRequest getAttribute("javax.servlet.forward.request_uri") returns the uri the user was on.

On tomcat, the getAttribute returns null, so to be backwards compatible, we must first check for the
attribute and if null, collect the uri from the function call.