Fixed checkstyle errors when running on Windows.
If I clone the project and build it on a Windows machine, I get a ton of NewlineAtEndOfFile errors.  This is because checkstyle is looking for a "CR-LF" at the end of each file on Windows, but only finding "LF" instead.  This fix tells checkstyle to be okay with only finding "LF".