Fixed a problem where the required LDAP Attributes regular expression error message does not show message to user.
When adding a Required LDAP Attribute to a forgotten password authentication, the Regular Expressions Error Message did not appear when the regular expression is not met.

Steps to reproduce:
Go to
Config Editor | Modules | Public | Forgotten Password | Profiles | default | Definition
Add a required LDAP Attribute and click Options.
Add a regular expression and a failure message.
Add the LDAP Attribute to one of the Optional Required authentication methods.
Try to use the forgotten password module as a user.
Input a response that does not match the regular expression.
Click "Check Answers"

Expected results:
The regular expression custom error message should appear on the screen in some way.

Actual results:
A bubble pops up with the text "Please match the requested format"