Added check to avoid sending token on SMS/Mail verification
If mail/SMS verification are enabled, the updateProfile process and the
newUser process will send a verification token regardless if the value
is empty or not. This would stop a user to complete the registration 
process in case email or mobile number are optional but verification
is enabled. The commit will skip sending the token if email/mobile
values are null or empty (even if the value is sent by the REST 
resolveTokenDestination client)