Style fixes to fix some display issues...
Changed the table headers so they stay on one line and don't wrap.  This way, the position of the first line of content is consistent across all the browsers (some browsers were putting the first line of content up into the header area).  This was seen when going to the Administration page, selecting Data Analysis, under the Directory Reporting tab, and Data Viewer sub-tab.

Updated the z-index of the plus icon, so it wasn't overlapping with the dialog content.  This was seen when going to the Configuration Manager, clicking Certificates, and clicking on one of the certificates in the list (the plus sign would show through the dialog content).