Fix misleading error message for Password_SameAsAttr
By default Pwm checks if the password contains cn, sn or givenName. When any of
these are allowed to be short (e.g. 2 characters), the likelihood of a long,
strong password consisting of random characters matching any of those is big. So
do not tell the user that password is "too obvious" if it happens to contain a
user attribute by accident. Rather give the user a better idea of the problem so
that he/she can fix it.